+++ .github/workflows/tests.yaml
@@ -77,7 +77,7 @@ jobs:
 
       - name: Install dask branch
         shell: bash -l {0}
-        run: python -m pip install --no-deps git+https://github.com/rjzamora/dask@nuke-hlg
+        run: python -m pip install --no-deps git+https://github.com/mrocklin/dask@nuke-hlg
 
       - name: Install
         shell: bash -l {0}
